Don't create FIFO file in scheduler

+++ docker/opensips_scheduler/opensips_scheduler.sh
@@ -4,6 +4,9 @@ set -e
 
 while :
 do
-  echo "::{\"jsonrpc\":\"2.0\",\"method\":\"lb_reload\"}" > $FIFO_NAME
+  if [ -f $FIFO_NAME ]; then
+    echo "::{\"jsonrpc\":\"2.0\",\"method\":\"lb_reload\"}" > $FIFO_NAME
+  fi
+
   sleep 30s
 done
